fix start position of consumers
Seeking at the requested start positions was not donei, depending on the
requested mode and actual consumer read position. Moreover, this was
not adjusted on rebalance. More work is needed to fully manage rebalacing.

Now read position offsets are always commited, this allows to measure
application lag of the consummers.

Also, no yield is done for each processed event. This was a big performances
bottleneck on pypy. A fixed countdown of 5000 event is used for now. This
probably needs to be reworked to ensure that other tasks can be scheduled even
on high workloads.